Refactor  e2e test and added new ones

+++ src/member/infrastructure/controller/member.controller.ts
@@ -28,14 +28,16 @@ import { MemberDto } from '../dto/member.dto';
 import { MemberNameAlreadyRegisteredError } from '../../domain/exception/member-name-in-group.error';
 import { MemberIdAlreadyRegisteredError } from '../../domain/exception/member-id-already-registered.error';
 import { MemberMadeTransactionError } from '../../domain/exception/member-made-transaction.error';
+import { GroupIdNotFoundError } from '../../../group/domain/exception/group-id-not-found.error';
+import { LastMemberInGroupError } from '../../domain/exception/last-member-in-group.error';
 
 @ApiTags('Members')
 @Controller('members')
 export class MemberController {
   constructor(private readonly memberService: MemberService) {}
 
-  @ApiOperation({ summary: 'Get Members of Member' })
-  @ApiResponse({ status: 204, description: 'Get Members of Member.' })
+  @ApiOperation({ summary: 'Get Members of Group' })
+  @ApiResponse({ status: 204, description: 'Get Members of Group.' })
   @ApiResponse({ status: 404, description: 'Not found' })
   @UseGuards(FirebaseAuthGuard)
   @Get(':id')
@@ -43,8 +45,8 @@ export class MemberController {
     try {
       return await this.memberService.getMembersByGroupId(params.id);
     } catch (e) {
-      if (e instanceof MemberIdNotFoundError) {
-        throw new NotFoundException('Member not found');
+      if (e instanceof GroupIdNotFoundError) {
+        throw new NotFoundException('Group not found');
       } else if (e instanceof Error) {
         throw new BadRequestException(`Unexpected error: ${e.message}`);
       } else {
@@ -69,6 +71,8 @@ export class MemberController {
     } catch (e) {
       if (e instanceof MemberIdAlreadyRegisteredError) {
         throw new ConflictException(e.message);
+      } else if (e instanceof GroupIdNotFoundError) {
+        throw new NotFoundException('Group Id not found');
       } else if (e instanceof MemberNameAlreadyRegisteredError) {
         throw new ConflictException(e.message);
       } else if (e instanceof Error) {
@@ -85,12 +89,20 @@ export class MemberController {
   @UseGuards(FirebaseAuthGuard)
   @HttpCode(204)
   @Delete(':id')
-  async removeMember(@Param() params) {
+  async removeMember(@Param() params, @Request() req) {
+    const idRequester: UserId = req.user;
+
+    if (idRequester.value === params.id) {
+      throw new ForbiddenException('Cannot delete requester');
+    }
+
     try {
       return await this.memberService.removeMember(params.id);
     } catch (e) {
       if (e instanceof MemberIdNotFoundError) {
         throw new NotFoundException('Member not found');
+      }else if (e instanceof LastMemberInGroupError) {
+        throw new NotAcceptableException(e.message);
       } else if (e instanceof MemberMadeTransactionError) {
         throw new BadRequestException(e.message);
       } else if (e instanceof Error) {
@@ -108,7 +120,7 @@ export class MemberController {
   @HttpCode(204)
   @Patch(':id')
   async registerMemberAsUser(
-    @Query('id') id: string,
+    @Param() params,
     @Body() memberDto: RegisterMemberAsUserDto,
     @Request() req,
   ) {
@@ -120,7 +132,7 @@ export class MemberController {
 
     try {
       return await this.memberService.registerMemberAsUser(
-        id,
+        params.id,
         memberDto.idUser,
       );
     } catch (e) {
